Modify RadioLib library to allow using illegal syncword 0x0F0F (to be able to communicate with Fossasat1

+++ lib/RadioLib/src/modules/SX126x/SX1268.h
@@ -47,7 +47,7 @@ class SX1268: public SX126x {
 
       \returns \ref status_codes
     */
-    int16_t begin(float freq = 434.0, float bw = 125.0, uint8_t sf = 9, uint8_t cr = 7, uint8_t syncWord = SX126X_SYNC_WORD_PRIVATE, int8_t power = 14, float currentLimit = 60.0, uint16_t preambleLength = 8, float tcxoVoltage = 1.6);
+    int16_t begin(float freq = 434.0, float bw = 125.0, uint8_t sf = 9, uint8_t cr = 7, uint16_t syncWord = SX126X_SYNC_WORD_PRIVATE, int8_t power = 14, float currentLimit = 60.0, uint16_t preambleLength = 8, float tcxoVoltage = 1.6);
 
     /*!
       \brief Initialization method for FSK modem.
